There are a ton of improvements here.
The documentation is more pertinent to the actual workflow, and now
includes instructions on how to apply the repo to a live master.

Changes to the bootstrap_r10k script to work a bit more cleanly.

Removed the @example.com from the master answers file for 3.7

Provision script turns off iptables on vagrant instances.

Removed resources in profile::puppet::master that clash with PE
3.7 internal management.

Updated vagrant box to centos 6.6 since 6.5 went bye-bye on the
puppetlabs vagrant account.