Update xmaster and xagent to official centos/7 box
This box should stay up to date, this box looks to be "official" on atlas:
https://atlas.hashicorp.com/centos/boxes/7 and have relatively frequently
releases.

It takes about 13 minutes to provision the xmaster with PE already downloaded.
This is on a MacBook Pro 13 2015.

There could be a potential issue with the yum install of the virtual box
additions, but this should only happen bringing the base box up.

Add note about vagrant-vbguest plugin which is required with these centos/7
boxes because there are no guest tools installed which means there is no way to
mount the shared folder to /vagrant.