Update master.txt answers for 2015.2.2
Without this patch, puppet enterprise fails to install with an error that it
cannot connect to the postgresql server:

    ==> xmaster: !! ERROR: Could not connect to the postgresql server using the
    ==> xmaster:    pe-postgres user.
    ==> xmaster:
    ==> xmaster: ========================================================================
    ==> xmaster: /vagrant/provision/bootstrap_r10k.sh: line 11: /opt/puppet/bin/puppet: No such file or directory
    ==> xmaster: /vagrant/provision/bootstrap_r10k.sh: line 15: /opt/puppet/bin/r10k: No such file or directory
    ==> xmaster: /vagrant/provision/bootstrap_r10k.sh: line 18: /opt/puppet/bin/puppet: No such file or directory
    ==> xmaster: All done! Now ssh in using vagrant ssh xmaster and sudo to root!

This addresses the problem by updating the answers file to 2015.2.2 format.
The file was generated by doing a manual installation of 2015.2.2 then
comparing the sorted answers to the 3.8 answers using `diff -U2`.

The actual change that resolved the above postgresql error is related to the
removal of these two lines, which were removed discretely, but squashed into
this commit. It's probably the `q_database_host` one that was causing the
issue with verification.

    q_database_host=localhost
    q_database_port=5432